Revive dependencies, fix random issues
