--- clippy.toml
@@ -1,1 +1,0 @@
-msrv = "1.85.0"
