+++ crates/smb-transport/Cargo.toml
@@ -13,9 +13,11 @@ readme.workspace = true
 
 [dependencies]
 smb-dtyp = { path = "../smb-dtyp" }
+smb-msg = { path = "../smb-msg", optional = true }
 
 log = { workspace = true }
 tokio = { workspace = true, optional = true }
+tokio-util = { workspace = true, optional = true }
 futures-core = { workspace = true, optional = true }
 futures-util = { workspace = true, optional = true }
 
@@ -25,7 +27,7 @@ rustls = { workspace = true, default-features = false, optional = true }
 rustls-platform-verifier = { workspace = true, default-features = false, optional = true }
 
 # RDMA
-async-rdma = { git = "https://github.com/AvivNaaman/async-rdma.git", rev = "b6f23d21806cfee5583b77f5f68f5e8184d50891", features = [
+async-rdma = { git = "https://github.com/AvivNaaman/async-rdma.git", rev = "878828dd", features = [
     "cm",
     "raw",
 ], optional = true }
@@ -38,9 +40,10 @@ maybe-async = { workspace = true }
 thiserror = { workspace = true }
 
 [features]
-default = []
-async = ["tokio", "futures-core", "futures-util"]
+default = ["netbios"]
+async = ["tokio", "tokio-util", "futures-core", "futures-util"]
 is_sync = ["maybe-async/is_sync"]
 
+netbios = []
 quic = ["dep:quinn", "dep:rustls", "dep:rustls-platform-verifier"]
-rdma = []
+rdma = ["dep:smb-msg", "dep:async-rdma"]
