+++ crates/smb-transport/scripts/ksmbd.sh
@@ -2,7 +2,7 @@
 modprobe rdma_rxe
 modprobe ib_uverbs
 modprobe rdma_cm
-rdma link add rxe_ens160 type rxe netdev ens160
+rdma link add rxe_ens18 type rxe netdev ens18
 # Start ksmbd
 modprobe ksmbd
 ksmbd.mountd
