+++ crates/smb-transport/scripts/ksmbd_kill.sh
