+++ crates/smb-transport/src/rdma/config.rs
@@ -1,0 +1,5 @@
+
+
+#[cfg(feature = "rdma")]
+#[derive(Debug, Default, Clone, PartialEq, Eq)]
+pub struct RdmaConfig {}
