+++ crates/smb-transport/src/rdma/smbd.rs
@@ -1,10 +1,13 @@
+#![allow(unused_parens)]
 ///! SMB-Direct (SMBD) packets & structures
 ///
 /// [MS-SMBD](https://learn.microsoft.com/en-us/openspecs/windows_protocols/ms-smbd/b25587c4-2507-47a4-aa89-e5d3f04f7197)
+
 use binrw::prelude::*;
 use modular_bitfield::prelude::*;
 
-use crate::packets::smb2::Status;
+use smb_msg::Status;
+
 
 const SMBD_VERSION: u16 = 0x100; // SMBD v1.0
