+++ crates/smb/src/lib.rs
@@ -41,7 +41,7 @@ pub use tree::{DfsRootTreeRef, Tree};
 pub use smb_dtyp::*;
 pub use smb_fscc::*;
 pub use smb_msg::*;
-pub use smb_transport::*;
+pub use smb_transport as transport;
 
 /// SMB Result type
 pub type Result<T> = std::result::Result<T, crate::Error>;
